Rearranged file structure for routes and created the guarded class.
The guarded class, different from the protected class previously, will now only check the authenticated app state and also check if the company_subdir matches the company subdir stored in the cookies. Ultimately this is just for the user client side experience, as eitherway the token is only valid for the help desk it was assigned from and is not able to be used in other help desks as the validation is conducted on the server.