Started creating the users area for the application.
Updated dashboardAppbar so it is now sticky and has a margin bottom of 40px.
Created a css class "standard-margin-bottom" which gives margin-bottom of 20px to keep spacing consistent throughout pages and easily manageable.
Created UsersTable module which loads the users from the api and shows them to screen. There is a edit button there is no functionality to it as of yet.