Updated view.js
Updated view.js to make setting analysts efficient, and also removed the error of unable to download users showing to the screen. Instead of showing empty as it would before, it checks if there was users downloaded or not, and if not, puts in the current analyst only so its never left empty.